+++ ReleaseNotes.md
@@ -1,5 +1,17 @@
 # Release Notes
 
+## Version 0.2.0
+
+### Features
+
+* Support of named credentials in Solace Connector
+  * up to 4 credentials per application 
+* Upgrade to Solace Connector v0.15.2 
+
+### Fixes
+
+* no bug fixes
+
 ## Version 0.1.0
 
 ### Features